Cleaned up a bunch of compiler warnings.
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@3240 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/fileloader.cpp
@@ -838,7 +838,7 @@ bool FileLoader::ReadPage(QString fileName, SCFonts &avail, ScribusDoc *doc, Scr
 			Its->BackBox = 0;
 			while (Its->NextIt != -1)
 			{
-				if (Its->NextIt < doc->Items.count())
+				if (Its->NextIt < static_cast<int>(doc->Items.count()))
 				{
 					Itn = doc->Items.at(Its->NextIt);
 					Its->NextBox = Itn;
